Clean types file
