--- src/types/categoriesEnum.ts
@@ -1,11 +1,0 @@
-enum CategoriesEnum {
-  Business = "business",
-  Entertainment = "entertainment",
-  General = "general",
-  Health = "health",
-  Science = "science",
-  Sports = "sports",
-  Technology = "technology",
-}
-
-export default CategoriesEnum;
