+++ src/types/index.ts
@@ -1,8 +1,8 @@
 export type { default as Article } from '@/types/article';
 export type { default as Source } from '@/types/source';
-export type { default as NewsApiResponse } from '@/types/newsApiResponse';
-export type { default as SearchInType } from '@/types/searchInType';
 export type { default as NewsListFilters } from '@/types/newsListFilters';
-export type { default as Categories } from '@/types/categories';
-export { default as CategoriesEnum } from '@/types/categoriesEnum';
-export { default as SearchInEnum } from '@/types/searchInEnum';
+export type { Categories } from '@/types/categories';
+export type { SearchInType } from '@/types/searchIn';
+// enums
+export { CategoriesEnum } from '@/types/categories';
+export { SearchInEnum } from '@/types/searchIn';
