--- src/types/newsApiResponse.ts
@@ -1,9 +1,0 @@
-import Article from "@/types/article";
-
-interface NewsApiResponse {
-  status: string;
-  totalResults: number;
-  articles: Article[];
-}
-
-export default NewsApiResponse;
