+++ src/types/searchIn.ts
@@ -1,0 +1,7 @@
+export enum SearchInEnum {
+  Title = 'title',
+  Description = 'description',
+  Content = 'content',
+}
+
+export type SearchInType = SearchInEnum.Title | SearchInEnum.Description | SearchInEnum.Content;
