--- src/types/searchInEnum.ts
@@ -1,7 +1,0 @@
-enum SearchInEnum {
-  Title = 'title',
-  Description = 'description',
-  Content = 'content',
-}
-
-export default SearchInEnum;
