--- src/types/searchInType.tsx
@@ -1,5 +1,0 @@
-import SearchInEnum from "@/types/searchInEnum";
-
-type SearchInType = SearchInEnum.Title | SearchInEnum.Description | SearchInEnum.Content;
-
-export default SearchInType;
